requirements.txt
streamlit==0.79.0
numpy==1.20.1
pandas==1.2.3
scipy==1.6.2
seaborn==0.11.1
matplotlib==3.4.1
Pillow==8.2.0
scikit_learn==0.24.1
